@@ -13,19 +13,19 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^14.0.6",
-    "@angular/cdk": "^14.0.5",
-    "@angular/common": "^14.0.6",
-    "@angular/compiler": "^14.0.6",
-    "@angular/core": "^14.0.6",
-    "@angular/forms": "^14.0.6",
-    "@angular/material": "^14.0.5",
-    "@angular/platform-browser": "^14.0.6",
-    "@angular/platform-browser-dynamic": "^14.0.6",
-    "@angular/router": "^14.0.6",
+    "@angular/animations": "^14.1.0",
+    "@angular/cdk": "^14.1.0",
+    "@angular/common": "^14.1.0",
+    "@angular/compiler": "^14.1.0",
+    "@angular/core": "^14.1.0",
+    "@angular/forms": "^14.1.0",
+    "@angular/material": "^14.1.0",
+    "@angular/platform-browser": "^14.1.0",
+    "@angular/platform-browser-dynamic": "^14.1.0",
+    "@angular/router": "^14.1.0",
     "rxjs": "^7.5.6",
     "tslib": "^2.4.0",
-    "zone.js": "^0.11.6"
+    "zone.js": "^0.11.7"
   },
   "devDependencies": {
     "@angular-devkit/build-angular": "^14.0.2",
