@@ -5,7 +5,8 @@ import { ReactiveFormsModule } from '@angular/forms';
 import { MaterialModule } from './material.module';
 import { RouterModule } from '@angular/router';
 import { AppRoutingModule } from './app-routing.module';
-import { HttpClientModule } from '@angular/common/http';
+import { HttpClientModule, HTTP_INTERCEPTORS } from '@angular/common/http';
+import { ErrorIntercept } from './error.interceptor';
 
 import { WeathercardComponent } from './components/weathercard/weathercard.component';
 import { FooterComponent } from './components/footer/footer.component';
@@ -14,7 +15,8 @@ import { SearchComponent } from './components/search/search.component';
 import { SearchpageComponent } from './pages/searchpage/searchpage.component';
 import { AppComponent } from './app.component';
 import { DetailsComponent } from './pages/details/details.component';
-
+import { DailyweatherComponent } from './components/dailyweather/dailyweather.component';
+import { TodayHighlightsComponent } from './components/today-highlights/today-highlights.component';
 @NgModule({
     imports: [
         BrowserAnimationsModule,
@@ -31,9 +33,14 @@ import { DetailsComponent } from './pages/details/details.component';
         HeaderComponent,
         FooterComponent,
         WeathercardComponent,
+        DailyweatherComponent,
         SearchpageComponent,
         DetailsComponent,
+        TodayHighlightsComponent,
     ],
     bootstrap: [AppComponent],
+    providers: [
+        { provide: HTTP_INTERCEPTORS, useClass: ErrorIntercept, multi: true },
+    ],
 })
 export class AppModule {}
