@@ -1,15 +1,17 @@
 import { EventEmitter, Injectable, Output } from '@angular/core';
-import { Observable, of } from 'rxjs';
+import { distinctUntilChanged, filter, Observable, of } from 'rxjs';
 import { switchMap, map } from 'rxjs';
 import { ApiService } from './api.service';
 @Injectable({
     providedIn: 'root',
 })
-export class InputCatchService {
+export class AutoComplete {
     constructor(private _api: ApiService) {}
 
-    getAutoCompleteArray(value: string): Observable<Array<string>> {
+    getAutoCompleteArray(value: string): Observable<string[]> {
         return of(value).pipe(
+            distinctUntilChanged(),
+            filter(val => val.length !== 0),
             switchMap((value: string) => this._api.getAutoComplete(value)),
             map((response: Observable<any>) => {
                 let arr: string[] = [];
